# conectatech-backend/docs/swagger/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ConectaTech API",
    "description": "API que conecta estudantes de diversos cursos da UNP a vagas de estagio na regiao de Mossoro/RN",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Registrar novo estudante",
        "description": "Cria uma nova conta de estudante na plataforma",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegistroRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Usuario registrado com sucesso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UsuarioDTO" }
              }
            }
          },
          "400": {
            "description": "Dados invalidos ou email ja cadastrado"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "description": "Autentica o usuario e retorna um token JWT",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "400": {
            "description": "Email ou senha invalidos"
          }
        }
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Solicitar recuperacao de senha",
        "description": "Gera um token de recuperacao para o email informado",
        "operationId": "forgotPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RecuperarSenhaRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Token de recuperacao gerado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Token de recuperacao gerado com sucesso" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Email nao encontrado"
          }
        }
      }
    },
    "/api/auth/reset-password": {
      "put": {
        "tags": ["Auth"],
        "summary": "Alterar senha",
        "description": "Altera a senha usando o token de recuperacao",
        "operationId": "resetPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AlterarSenhaRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Senha alterada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Senha alterada com sucesso" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Token invalido ou expirado"
          }
        }
      }
    },
...
    "/api/jobs": {
      "get": {
        "tags": ["Vagas"],
        "summary": "Listar vagas",
        "description": "Lista todas as vagas com filtros opcionais por texto, localizacao e fonte",
        "operationId": "listJobs",
        "parameters": [
          {
            "name": "busca",
            "in": "query",
            "description": "Texto para buscar em titulo, empresa e descricao",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "localizacao",
            "in": "query",
            "description": "Filtrar por localizacao (ex: Mossoro)",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "fonte",
            "in": "query",
            "description": "Filtrar por fonte: JOOBLE, JSOUP ou ADMIN",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de vagas retornada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/VagaDTO" }
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{id}": {
      "get": {
        "tags": ["Vagas"],
        "summary": "Buscar vaga por ID",
        "description": "Retorna os detalhes de uma vaga especifica",
        "operationId": "getJobById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da vaga",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "Vaga encontrada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VagaDTO" }
              }
            }
          },
          "404": {
            "description": "Vaga nao encontrada"
          }
        }
      }
    },
    "/api/jobs/search-external": {
      "get": {
        "tags": ["Vagas"],
        "summary": "Buscar vagas externas",
        "description": "Importa vagas do Jooble e retorna resultados do banco",
        "operationId": "searchExternal",
        "parameters": [
          {
            "name": "keywords",
            "in": "query",
            "description": "Palavras-chave para busca",
            "required": false,
            "schema": { "type": "string", "default": "estagio" }
          },
          {
            "name": "location",
            "in": "query",
            "description": "Localizacao da busca",
            "required": false,
            "schema": { "type": "string", "default": "Mossoro, RN" }
          }
        ],
        "responses": {
          "200": {
            "description": "Vagas externas importadas e retornadas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/VagaDTO" }
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/select": {
      "post": {
        "tags": ["Vagas"],
        "summary": "Selecionar vaga",
        "description": "Adiciona uma vaga a lista de vagas selecionadas do estudante",
        "operationId": "selectJob",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SelecionarVagaRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Vaga selecionada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Vaga selecionada com sucesso" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token JWT ausente ou invalido"
          },
          "404": {
            "description": "Vaga nao encontrada"
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/jobs/my-selections": {
      "get": {
        "tags": ["Vagas"],
        "summary": "Minhas vagas",
        "description": "Lista as vagas selecionadas pelo estudante autenticado",
        "operationId": "mySelections",
        "responses": {
          "200": {
            "description": "Lista de vagas selecionadas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/VagaDTO" }
                }
              }
            }
          },
          "401": {
            "description": "Token JWT ausente ou invalido"
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/jobs/my-selections/{vagaId}": {
      "delete": {
        "tags": ["Vagas"],
        "summary": "Remover vaga selecionada",
        "description": "Remove uma vaga da lista de selecionadas do estudante",
        "operationId": "removeSelection",
        "parameters": [
          {
            "name": "vagaId",
            "in": "path",
            "description": "ID da vaga",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "Vaga removida das selecionadas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Vaga removida das selecionadas" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token JWT ausente ou invalido"
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
      "RegistroRequest": {
        "required": ["nome", "email", "senha"],
        "type": "object",
        "properties": {
          "nome": { "type": "string", "description": "Nome completo", "example": "Joao Silva" },
          "email": { "type": "string", "description": "Email do estudante", "example": "joao.silva@email.com" },
          "senha": { "type": "string", "description": "Senha (min 6 caracteres)", "minLength": 6, "example": "minhasenha123" },
          "curso": { "type": "string", "description": "Curso do estudante", "example": "Ciencia da Computacao" },
          "periodo": { "type": "string", "description": "Periodo atual", "example": "6" }
        },
        "description": "Dados para registro de novo estudante"
      },
      "LoginRequest": {
        "required": ["email", "senha"],
        "type": "object",
        "properties": {
          "email": { "type": "string", "description": "Email do usuario", "example": "aluno@email.com" },
          "senha": { "type": "string", "description": "Senha do usuario", "example": "minhasenha123" }
        },
        "description": "Dados para login"
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string", "description": "Token JWT para autenticacao", "example": "eyJhbGciOiJIUzI1NiJ9..." },
          "nome": { "type": "string", "description": "Nome do usuario", "example": "Joao Silva" },
          "email": { "type": "string", "description": "Email do usuario", "example": "aluno@email.com" },
          "role": { "type": "string", "description": "Role do usuario", "example": "STUDENT" }
        },
        "description": "Resposta de login com token JWT"
      },
      "RecuperarSenhaRequest": {
        "required": ["email"],
        "type": "object",
        "properties": {
          "email": { "type": "string", "description": "Email cadastrado", "example": "aluno@email.com" }
        },
        "description": "Dados para solicitar recuperacao de senha"
      },
      "AlterarSenhaRequest": {
        "required": ["token", "novaSenha"],
        "type": "object",
        "properties": {
          "token": { "type": "string", "description": "Token recebido na recuperacao de senha", "example": "550e8400-e29b-41d4-a716-446655440000" },
          "novaSenha": { "type": "string", "description": "Nova senha (min 6 caracteres)", "minLength": 6, "example": "novasenha456" }
        },
        "description": "Dados para alterar senha com token de recuperacao"
      },
...
      "UsuarioDTO": {
        "required": ["nome"],
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64", "description": "ID do usuario", "example": 1 },
          "nome": { "type": "string", "description": "Nome completo", "example": "Joao Silva" },
          "email": { "type": "string", "description": "Email", "example": "joao.silva@email.com" },
          "curso": { "type": "string", "description": "Curso", "example": "Ciencia da Computacao" },
          "periodo": { "type": "string", "description": "Periodo", "example": "6" },
          "role": { "type": "string", "description": "Role do usuario", "example": "STUDENT" },
          "criadoEm": { "type": "string", "description": "Data de criacao", "example": "30/03/2026 14:30" }
        },
        "description": "Dados do usuario"
      },
      "VagaDTO": {
        "required": ["titulo", "empresa"],
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64", "description": "ID da vaga", "example": 1 },
          "titulo": { "type": "string", "description": "Titulo da vaga", "example": "Estagiario de Desenvolvimento Web" },
          "empresa": { "type": "string", "description": "Nome da empresa", "example": "Tech Solutions LTDA" },
          "descricao": { "type": "string", "description": "Descricao da vaga", "example": "Vaga para desenvolvimento web com React e Spring Boot" },
          "localizacao": { "type": "string", "description": "Localizacao da vaga", "example": "Mossoro, RN" },
          "url": { "type": "string", "description": "URL da vaga original", "example": "https://exemplo.com/vaga/123" },
          "fonte": { "type": "string", "description": "Fonte da vaga", "example": "JOOBLE" },
          "dataPublicacao": { "type": "string", "description": "Data de publicacao", "example": "30/03/2026 10:00" }
        },
        "description": "Dados de uma vaga de estagio"
      },
      "SelecionarVagaRequest": {
        "required": ["vagaId"],
        "type": "object",
        "properties": {
          "vagaId": { "type": "integer", "format": "int64", "description": "ID da vaga a ser selecionada", "example": 1 }
        },
        "description": "Dados para selecionar uma vaga"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Token JWT obtained via POST /api/auth/login",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}